test(core): migrate UpdateTabTodoTests (Wave D)
32 tests, 74 failure sites preserved 1:1.
